feat(validate): add multiple file arguments to CLI
The CLI can now accept multiple file arguments. These arguments can be
file paths or glob-pattern. In case of glob-pattern, all the matched
paths are added to validation.

For `bundle` command it
explicitly checks that only one path is given. If multiple arguments are
given of a glob-pattern which matches multiple files it
throws an error.